Confirm deletion of files, add link to open astrometry image in new tab, sorting on manage files page

+++ analysis/utils/general.py
@@ -104,10 +104,10 @@ def delete_folders(fits_file):
     """
 
     # Remove all folders we drop during analysis
-    if os.path.exists(os.path.join(settings.CATALOGUE_ORIGINAL_DIRECTORY, fits_file.catalogue_filename)):
+    if os.path.exists(os.path.join(settings.CATALOGUE_ORIGINAL_DIRECTORY, fits_file.catalogue_filename)) and (fits_file.catalogue_filename is not (None, '')):
         os.remove(os.path.join(settings.CATALOGUE_ORIGINAL_DIRECTORY, fits_file.catalogue_filename))
 
-    if os.path.exists(os.path.join(settings.CATALOGUE_PROCESSED_DIRECTORY, fits_file.catalogue_filename)):
+    if os.path.exists(os.path.join(settings.CATALOGUE_PROCESSED_DIRECTORY, fits_file.catalogue_filename)) and (fits_file.catalogue_filename is not (None, '')):
         os.remove(os.path.join(settings.CATALOGUE_PROCESSED_DIRECTORY, fits_file.catalogue_filename))
 
     if os.path.exists(os.path.join(settings.FITS_DIRECTORY, fits_file.fits_filename)):
